testing Float32 speed

+++ src/common.jl
@@ -1,5 +1,5 @@
 
-function calculate_px(x::Array{Float64,2})
+function calculate_px(x::Array{Float32,2})
     XtX = transpose(x)*x
     result = X*inv(XtX)*transpose(x)
     # display(result)
@@ -15,20 +15,20 @@ function process_x(x)
     #         x[row] = -1.0
     #     end
     # end
-    # return convert(Array{Float64,1}, x)
+    # return convert(Array{Float32,1}, x)
 
     return rand([-1.0,1.0], size(x)[1])
 end
 
 function get_standardized_matrix(m)
     for col in 1:size(m)[2]
-        summ::Float64 = 0.0
+        summ::Float32 = 0.0
         rows = size(m)[1]
         for row in 1:rows
             summ += m[row, col]
         end
         mean = summ/rows
-        sums::Float64 = 0.0
+        sums::Float32 = 0.0
         for row in 1:rows
             sums += (m[row,col] - mean)^2
         end
